added user pending function
changed the database to have api keys for ervaringsdeskundige and organisatie.  made it so whenever someone registers a account the accepteerd column is set to false automatically. changed the style to fit the template. added jinja for the header and footer for both the admin and user side. Made an extra login button specifically for ervaringsdeskundige. made it so the pending page so the admin can approve or disapprove new registered accounts. and fixed the top right session type and session name.