added small updates and failsafes
added login and register errors for wrong credentials and already existing email address when registering. added a search and filter for the accounts page. fixed session name not showing on the accounts page and edit account page. fixed register not working if supervisor checkbox wasnt checked. made it so ervaringsdeskundige only need email and api to login instead of phonenumber , api and email. added extra styling. added a "show details" button for pending accounts, to show all of the information they filled in instead of only being able to see their name, email and phonenumber.